Vidas y collder bajo los pinchos

+++ Gloton Galactico/Game/Source/Player.h
@@ -36,7 +36,7 @@ public:
 	
 	bool godMode;
 
-	int vidas;
+	float vidas;
 
 private:
